# src/config/swagger.json
{
  "openapi": "3.0.0",
  "info": {
    "title": "API Vizion",
    "description": "Description Example",
    "version": "1.0.0"
  },
  "servers": [
    {
      "url": "https://creepy-top-hat-bull.cyclic.cloud/",
      "description": "Online server"
    },
    {
      "url": "http://localhost:3000/",
      "description": "Local server"
    }
  ],
  "paths": {
...
    "/lead/{codigo}":{
      "get": {
        "tags": [
          "Lead"
        ],
        "description": "retorna um registro especifico pelo codigo",
        "operationId": "findByCodigo",
        "parameters": [
          {
            "name": "codigo",
            "in": "path",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "patch": {
        "tags": ["Lead"],
        "description": "Atualiza um registro especifico",
        "operationId": "updateByCodigo",
        "parameters": [
          {
            "name": "codigo",
            "in": "path",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LeadSemCodigo"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      },
      "delete": {
        "tags": ["Lead"],
        "description": "Remove um registro especifico",
        "operationId": "deleteByCodigo",
        "parameters": [
          {
            "name": "codigo",
            "in": "path",
            "required": true,
            "style": "simple",
            "explode": false,
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Success"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "LeadSemCodigo": {
        "required": [
          "telefone"
        ],
        "type": "object",
        "properties": {
          "nomeCompleto": {
            "type": "string"
          },
          "resposavel": {
            "type": "string"
          },
          "tipoLead": {
            "type": "string"
          },
          "empresa": {
            "type": "array",
            "items": [{
              "type": "object",
              "properties": {
                "name": {
                  "type": "string"
                },
                "value": {
                  "type": "string"
                }
              }
            }]
          },
          "origem": {
            "type": "array",
            "items": [{
              "type": "object",
              "properties": {
                "name": {
                  "type": "string"
                },
                "value": {
                  "type": "string"
                }
              }
            }]
          },
          "dataNascimento": {
            "type": "string",
            "format": "Date"
          },
          "cpf": {
            "type": "string"
          },
          "nacionalidade": {
            "type": "string"
          },
          "telefone": {
            "type": "string"
          },
          "email": {
            "type": "string"
          },
          "website": {
            "type": "string"
          },
          "endereco": {
            "type": "object",
            "properties": {
              "cep": {
                "type": "string"
              },
              "logradouro": {
                "type": "string"
              },
              "numero": {
                "type": "string"
              },
              "bairro": {
                "type": "string"
              },
              "cidade": {
                "type": "string"
              },
              "estado": {
                "type": "string"
              },
              "pais": {
                "type": "string"
              }
            }
          },
          "status": {
            "type": "string"
          },
          "companhia": {
            "type": "string"
          },
          "campanha_n": {
            "type": "array",
            "items": [{
              "type": "object",
              "properties": {
                "codigo": {
                  "type": "string"
                }
              }
            }
            ]
          },
          "criadoPor": {
            "type": "string"
          },
          "atualizadoPor": {
            "type": "string"
          }
        }
      },
...
    }
  }
}